툴킷 버전 업 0.0.5
- 타임스탬프 관련 핼퍼 함수 추가
- auth 형식 long 으로 변경
- 툴킷 버전 변경 0.0.5

+++ packages/io.witchcompany.toolkit/Editor/Scripts/DataStructure/Json/JAuth.cs
@@ -5,9 +5,9 @@ namespace WitchCompany.Toolkit.Editor.DataStructure
     [System.Serializable]
     public class JAuth
     {
-         public string accessToken;
-        public int accessExpire;
+        public string accessToken;
+        public long accessExpire;
         public string refreshToken;
-        public int refreshExpire;
+        public long refreshExpire;
     }
 }
